refactor home page background and remove unused home files
