+++ src/components/Home/Landing/Hero.module.css
@@ -1,0 +1,50 @@
+.hero_container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+
+    height: 100%;
+
+    border-radius: var(--border-radius);
+
+    background:  linear-gradient(180deg, var(--body-bg) 0%, var(--body-bg) 0%, rgba(0, 0, 0, 0) 0%, var(--body-bg) 100%), url('../../../assets/images/home/home_wallpaper.png') no-repeat
+        center center;
+
+    -webkit-background-size: cover;
+    -moz-background-size: cover;
+    -o-background-size: cover;
+    background-size: cover;
+}
+.hero_container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 2rem;
+}
+
+.hero_container img {
+    width: 250px;
+    transition: all var(--animation-speed) ease-in-out;
+}
+.ambient_background {
+    display: none;
+}
+.ambient {
+    background: var(--title-gradient);
+    font-family: var(--font-logo);
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+    background-clip: text;
+
+    font-size: clamp(3rem, 12vw + 1rem, 8rem);
+    font-weight: 100;
+}
+
+
+@media only screen and (min-width: 768px) {
+    .hero_container img {
+        width: 500px;
+    }
+}
