--- src/components/Home/Landing/Home1.module.css
@@ -1,133 +1,0 @@
-.slide_container {
-    background: linear-gradient(
-        180deg,
-        rgba(13, 17, 23, 0) 0%,
-        var(--dark1) 88.54%
-    );
-    /* background: darkblue; */
-    padding-bottom: 1rem;
-}
-
-.content_container {
-    width: 100%;
-    height: 100%;
-
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 2rem;
-}
-
-.content_container h1,
-.faster_section h1 {
-    font-family: var(--font-logo);
-    color: var(--text1);
-    font-weight: 200;
-    font-size: 34px;
-    line-height: 50px;
-    text-align: center;
-}
-
-.row_container {
-    /* padding-top: 50px; */
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 2rem;
-}
-
-/* .row_container img{
-    width: 331px;
-height: 420px;
-} */
-
-.column_reverse {
-    flex-direction: column-reverse;
-}
-.row_container section {
-    max-width: 907px;
-    display: flex;
-    flex-direction: column;
-    gap: 2rem;
-    align-items: center;
-    text-align: center;
-}
-
-.row_container h2 {
-    font-family: var(--font-logo);
-    color: var(--text1);
-    font-weight: 200;
-    font-size: 30px;
-    line-height: 38px;
-}
-
-.row_container p {
-    font-weight: 200;
-    font-size: 18px;
-    line-height: 22px;
-    color: var(--text1);
-}
-
-@media only screen and (min-width: 768px) {
-    .slide_container {
-        padding-bottom: 0;
-        padding: 2rem 0;
-        justify-content: center;
-        align-items: center;
-        display: flex;
-    }
-
-    .height_large {
-        min-height: 500px;
-    }
-
-    .height_small {
-        height: 300px;
-    }
-
-    .height_medium {
-        height: 400px;
-    }
-    .row_container {
-        flex-direction: row;
-    }
-    .content_container h1 {
-        text-align: start;
-    }
-}
-
-
-.home2{
-
-    background: url('../../../assets/images/home/home2.png') no-repeat
-        center center;
-
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-}
-.home3{
-
-    background: url('../../../assets/images/home/home3.png') no-repeat
-        center center;
-
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-}
-.home4{
-
-    background: url('../../../assets/images/home/home4.png') no-repeat
-        center center;
-
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-    border-bottom: 1px solid var(--border);
-
-}
