--- src/components/Home/Landing/HomeSlider.module.css
@@ -1,4 +1,0 @@
-.container .myswiper {
-    width: 100%;
-    height: 480px;
-}
