--- src/components/Home/Landing/Landing.module.css
@@ -1,17 +1,0 @@
-.main_container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-
-    height: 100%;
-
-    border-radius: var(--border-radius);
-    background: url('../../../assets/images/home/featured.png') no-repeat center
-        center;
-
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-}
