--- src/components/Home/Landing/Landing1.module.css
@@ -1,113 +1,0 @@
-.main_container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-
-    height: 100%;
-
-    border-radius: var(--border-radius);
-
-    background: url('../../../assets/images/home/home1.png') no-repeat center
-        bottom;
-
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-}
-.content_container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 2rem;
-}
-
-.content_container img {
-    width: 250px;
-    transition: all var(--animation-speed) ease-in-out;
-}
-.ambient_background {
-    display: none;
-}
-.ambient {
-    background: var(--title-gradient);
-    font-family: var(--font-logo);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-
-    font-size: clamp(3rem, 12vw + 1rem, 8rem);
-    font-weight: 100;
-}
-
-.action_button {
-    width: 100%;
-    z-index: 2;
-
-    max-width: 16rem;
-    padding: 8px 16px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 16px;
-
-    line-height: 24px;
-
-    max-height: 3rem;
-    cursor: pointer;
-    text-align: center;
-
-    transition: all var(--animation-speed) ease-in-out;
-    background: var(--highlight-gradient);
-    background-position: 1% 50%;
-    background-size: 300% 300%;
-    text-decoration: none;
-    border: 1px solid var(--dark3);
-    /* animation: shine 8s linear infinite; */
-    /* color: var(--accent5); */
-}
-
-.button_text {
-    font-family: var(--font-logo);
-}
-
-@keyframes shine {
-    0% {
-        background-position: 0% 50%;
-    }
-    50% {
-        background-position: 100% 50%;
-        color: rgb(0, 0, 0);
-    }
-    100% {
-        background-position: 0% 50%;
-        color: rgb(0, 0, 0);
-    }
-}
-
-.action_button:hover {
-    border: 1px solid rgba(223, 190, 106, 0);
-    background-position: 99% 50%;
-    /* color: black; */
-    animation-play-state: paused;
-}
-.action_button p {
-    color: var(--accent5);
-    font-weight: 100;
-    font-size: 30px;
-    line-height: 38px;
-}
-
-.action_button p:hover {
-    color: var(--dark2);
-
-    /* color: var(--dark2); */
-}
-
-@media only screen and (min-width: 768px) {
-    .content_container img {
-        width: 500px;
-    }
-}
