--- src/components/Home/Landing/Landing2.module.css
@@ -1,85 +1,0 @@
-.main_container {
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    align-items: center;
-    width: 100%;
-
-    height: 100%;
-
-    border-radius: var(--border-radius);
-
-    background: url('../../../assets/images/home/landing2.png') no-repeat center
-        center;
-
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-    gap: 8px;
-}
-.content_container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    /* gap: 1rem; */
-}
-.on_chain {
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-    font-family: var(--font-logo);
-    font-weight: 200;
-    font-size: 18px;
-    line-height: 22px;
-    color: var(--accent5);
-}
-.architecture {
-    width: 100%;
-
-    display: flex;
-    flex-direction: column;
-
-    text-align: center;
-    color: var(--text2);
-    font-weight: 200;
-    font-size: 18px;
-    line-height: 22px;
-}
-
-.highlight_text {
-    font-weight: 500;
-    color: var(--accent5);
-}
-
-.ambient {
-    background: var(--title-gradient);
-    font-family: var(--font-logo);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-
-    font-size: clamp(3rem, 12vw + 1rem, 8rem);
-    font-weight: 100;
-
-    display: none;
-}
-
-/* Desktop screen Sizes */
-@media only screen and (min-width: 668px) {
-    .ambient {
-        display: flex;
-    }
-
-    .highlight_text {
-        margin: 0 5px;
-    }
-    .architecture,
-    .on_chain {
-        flex-direction: row;
-        font-size: 24px;
-        line-height: 30px;
-    }
-}
