--- src/components/Home/Landing/Landing3.module.css
@@ -1,202 +1,0 @@
-.main_container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    width: 100%;
-
-    height: 100%;
-
-    border-radius: var(--border-radius);
-
-    background: url('../../../assets/images/home/landing3.png') no-repeat center
-        center;
-
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-    gap: 8px;
-}
-
-.content_container {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    gap: 10px;
-}
-.header_content {
-    display: flex;
-    flex-direction: column;
-    font-size: 24px;
-    line-height: 30px;
-    text-align: center;
-
-    font-family: var(--font-logo);
-}
-
-.header_text {
-    color: var(--accent5);
-    font-weight: 100;
-}
-
-.ambient_text {
-    background: var(--title-gradient);
-    font-family: var(--font-logo);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-}
-
-.highlight_text {
-    color: var(--accent5);
-    font-weight: 300;
-}
-.amm_pools {
-    font-size: 16px;
-    color: var(--text2);
-}
-.efficiency {
-    font-weight: 300;
-    font-size: 16px;
-    color: var(--accent5);
-}
-
-/* -------------------------- */
-
-.multi_contract_container,
-.single_contract_container {
-    width: 450px;
-    height: 333px;
-    display: flex;
-    flex-direction: column;
-    /* align-items: center; */
-    padding: 16px;
-    gap: 8px;
-    background: var(--dark2);
-    backdrop-filter: blur(10px);
-
-    border-radius: 4px;
-}
-
-.multi_contract_container h2,
-.single_contract_container h2 {
-    font-weight: 300;
-    font-size: 16px;
-    line-height: 20px;
-    color: var(--accent5);
-}
-.multi_contract_container h3,
-.single_contract_container h3 {
-    font-weight: 100;
-    font-size: 12px;
-    line-height: 15px;
-    color: var(--text2);
-}
-.multi_contract_circles {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-.multi_contract_circles .row {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-}
-.pool_display_container {
-    display: flex;
-    flex-direction: column;
-
-    border-radius: 50%;
-    border: 0.5px solid var(--accent5);
-    background: var(--dark1);
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.multi_contract_container .pool_display_container {
-    width: 60.83px;
-    height: 60.83px;
-}
-.pool_display_container p {
-    font-weight: 400;
-    font-size: 8px;
-    line-height: 10px;
-    color: var(--accent5);
-}
-
-.wallet_text {
-    width: 104.75px;
-    height: 104.75px;
-    border-radius: 50%;
-    border: 0.5px solid var(--accent5);
-    background: var(--dark1);
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.single_contract_circles {
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-    align-items: center;
-}
-.wallet_text p {
-    font-weight: 400;
-    font-size: 12px;
-    line-height: 15px;
-    color: var(--accent5);
-}
-
-.single_contract_container .pool_display_container {
-    width: 53.19px;
-    height: 53.19px;
-}
-.single_contract_container .row {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.pool_circle_container {
-    background: var(--title-gradient);
-    padding: 1px;
-    width: 140px;
-    height: 140px;
-    border-radius: 50%;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.pool_circle_content {
-    width: 100%;
-    height: 100%;
-    border-radius: 50%;
-
-    background: var(--dark2);
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 4px;
-}
-.single_contract_container,
-.multi_contract_container {
-    display: none;
-}
-
-/* Desktop screen Sizes */
-@media only screen and (min-width: 768px) {
-    .single_contract_container,
-    .multi_contract_container {
-        display: flex;
-    }
-}
