--- src/components/Home/Landing/Landing4.module.css
@@ -1,129 +1,0 @@
-.main_container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    gap: 10px;
-
-    height: 100%;
-
-    border-radius: var(--border-radius);
-
-    background: url('../../../assets/images/home/landing4.png') no-repeat center
-        center;
-
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-    font-family: var(--font-logo);
-}
-
-.content_container {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-}
-
-.content_container img {
-    display: none;
-}
-
-.header_content {
-    display: flex;
-    flex-direction: column;
-    font-size: 24px;
-    line-height: 30px;
-    text-align: center;
-
-    font-family: var(--font-logo);
-}
-
-.header_text {
-    color: var(--accent5);
-    font-weight: 100;
-}
-
-.ambient_text {
-    background: var(--title-gradient);
-    font-family: var(--font-logo);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-}
-
-.highlight_text {
-    color: var(--accent5);
-    font-weight: 300;
-    margin: 0 5px;
-}
-.amm_pools {
-    font-size: 16px;
-    color: var(--text2);
-}
-
-.row {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 10px;
-}
-
-.liquidity_content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    padding: 8px 16px;
-    gap: 8px;
-
-    background: var(--dark2);
-    backdrop-filter: blur(20px);
-
-    border-radius: var(--border-radius);
-}
-
-.half_content {
-    text-align: center;
-}
-.liquidity_content img {
-    max-width: 100%;
-}
-
-.content_title {
-    font-weight: 300;
-    font-size: 16px;
-    line-height: 20px;
-    color: var(--accent5);
-}
-.content_text {
-    font-weight: 100;
-    font-size: 12px;
-    line-height: 15px;
-    text-align: center;
-    color: var(--text2);
-    word-wrap: break-word;
-}
-
-@media only screen and (min-width: 768px) {
-    .row {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        gap: 10px;
-    }
-
-    .content_container img {
-        display: flex;
-    }
-    .half_content {
-        width: 442.14px;
-        height: 177px;
-        text-align: justify;
-    }
-
-    .content_text {
-        text-align: justify;
-    }
-}
