--- src/components/Home/Landing/Landing5.module.css
@@ -1,27 +1,0 @@
-.main_container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-
-    height: 100%;
-
-    border-radius: var(--border-radius);
-
-    background: url('../../../assets/images/home/landing6.png') no-repeat center
-        center;
-
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-
-    font-family: var(--font-logo);
-    font-style: normal;
-    font-weight: 100;
-    font-size: 24px;
-    line-height: 30px;
-    color: var(--accent5);
-
-    text-align: center;
-}
