+++ src/components/Home/Stats/Stats.module.css
@@ -3,7 +3,7 @@
     flex-direction: column;
 
     gap: 1rem;
-    padding: 4rem 0;
+    padding: 1rem 0;
 }
 .title {
     font-style: normal;
