+++ src/components/Home/TopPools/TopPools.module.css
@@ -33,8 +33,4 @@
     gap: 1rem;
 }
 
-.divider {
-    width: 100%;
-    height: 1px;
-    background: var(--border);
-}
+
