+++ src/i18n/en/common.json
@@ -30,6 +30,6 @@
     "slide6": {
         "part1": "something something gas savings..."
     },
-    "homeStatsTitle": "DEX Stats",
+    "homeStatsTitle": "Ambient Finance Stats",
     "ourInvestorsTitle": "Our Investors"
 }
