+++ src/pages/Home/Home.tsx
@@ -1,21 +1,22 @@
 import DividerDark from '../../components/Global/DividerDark/DividerDark';
-import HomeSlider from '../../components/Home/Landing/HomeSlider';
 import Stats from '../../components/Home/Stats/AmbientStats';
 import TopPools from '../../components/Home/TopPools/TopPools';
 import styles from './Home.module.css';
-import Home1 from '../../components/Home/Landing/Home1';
+import LandingSections from '../../components/Home/Landing/LandingSections';
+import Hero from '../../components/Home/Landing/Hero';
 
 export default function Home() {
     return (
         <section data-testid={'home'} className={styles.home_container}>
-            <HomeSlider />
+            <div style={{ width: '100%', height: '480px' }}>
+                <Hero />
+            </div>
             <div className={styles.pools_container}>
                 <TopPools />
-                <DividerDark />
                 <Stats />
             </div>
             <DividerDark />
-            <Home1 />
+            <LandingSections />
         </section>
     );
 }
